+++ src/constants/index.js
@@ -12,6 +12,7 @@ export const APP_NAME = 'Markdown notes';
 
 
 export const USER_SESSION_KEY = 'verida_user_loggedin';
+export const VERIDA_USER_SIGNATURE = '_verida_auth_user_signature';
 
 export let DATASTORE_SCHEMA;
